requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
web3==6.10.0
eth-account==0.8.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.4
pandas==2.1.1
joblib==1.3.2
Pillow==10.1.0
scikit-learn==1.3.1
tensorflow==2.14.0
qrcode==7.4.2
razorpay==1.4.2